PkgBump

Documentation for PkgBump.

PkgBump.bump

PkgBump.bumpmajor

PkgBump.bumpminor

PkgBump.bumppatch

PkgBump.updatemajor

PkgBump.updatemajor!

PkgBump.updateminor

PkgBump.updateminor!

PkgBump.updatepatch

PkgBump.updatepatch!

PkgBump.updateversion

PkgBump.updateversion!

PkgBump.bump — Method

bump(mode::Symbol; commit=true, push=true)

Bumps the version of the current active project according to mode, commits the change to a new branch, and pushes the branch to the remote repository.

PkgBump.bumpmajor — Method

bumpmajor(;kwargs)

Bump the major version of the current active project, commit, and push the changes.

PkgBump.bumpminor — Method

bumpminor(;kwargs)

Bump the minor version of the current active project, commit, and push the changes.

PkgBump.bumppatch — Method

bumppatch(;kwargs)

Bump the patch version of the current active project, commit, and push the changes.

PkgBump.updatemajor! — Method

updatemajor!(project::Pkg.Types.Project, project_file)

Increment the major version of the given project and write the changes to the project_file.

PkgBump.updatemajor — Method

updatemajor(project_file::AbstractString)

Update the major version of the project defined in project_file.

PkgBump.updateminor! — Method

updateminor!(project::Pkg.Types.Project, project_file::AbstractString)

Increment the minor version of the given project and write the changes to the project_file.

PkgBump.updateminor — Method

updateminor(project_file::AbstractString)

Update the minor version of the project defined in project_file.

PkgBump.updatepatch! — Method

updatepatch!(project::Pkg.Types.Project, project_file::AbstractString)

Increment the patch version of the given project and write the changes to the project_file.

PkgBump.updatepatch — Method

updatepatch(project_file::AbstractString)

Update the patch version of the project defined in project_file.

PkgBump.updateversion! — Method

updateversion!(project::Pkg.Types.Project, project_file::AbstractString, mode::Symbol)

Update the version of the given project file according to the specified mode (:patch, :minor, :major). The new version is written directly to the project_file.

PkgBump.updateversion — Method

updateversion(project_file::AbstractString, mode::Symbol) -> Pkg.Types.Project

Read the project from project_file, update its version according to mode, and write the changes back. Returns the updated project.